课程组队模块初步完成

+++ src/main/java/com/example/scd/dao/CharacterDao.java
@@ -1,5 +1,5 @@
 package com.example.scd.dao;
 
 public interface CharacterDao {
-    Integer getNumOfCharacter();
+    Long getNumOfCharacter();
 }
